Extracted `ConvertingGenericData` to a standalone class; Tidying up;
